requirements.txt
python-nmap==0.7.1
sqlalchemy==1.4.46
psutil==5.9.4
flask==2.3.3
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0
requests==2.28.2
schedule==1.1.0
gunicorn==20.1.0
werkzeug==2.3.7
# Cloud providers
boto3==1.28.38
azure-identity==1.13.0
azure-mgmt-resource==23.0.1
google-cloud-compute==1.12.0
oci==2.104.1
